requirements.txt
beautifulsoup4
st-pages
openai
pysqlite3-binary
chromadb==0.5.16
langchain_chroma
langchain_community
langchain-openai
langchain_text_splitters
langchain_community
streamlit_timeline
sentence-transformers
langchain_huggingface
streamlit-lottie